🚚 rename: 메인페이지 폴더로 병합 #110
